fix: change important params

+++ src/database/models.py
@@ -7,7 +7,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = "users"
 
-    id = sa.Column(sa.INTEGER, unique=True, nullable=False, primary_key=True)
+    id = sa.Column(sa.BIGINT, unique=True, nullable=False, primary_key=True)
 
     def __str__(self) -> str:
         return f"<User: {self.id}>"
